new/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15,<4.0
channels>=4.0,<5.0
psycopg[binary]>=3.1
python-dotenv>=1.0
Pillow>=10.3
reportlab>=4.0
img2pdf>=0.5
pdf2image>=1.17
uvicorn[standard]>=0.30
structlog>=24.1
pydantic>=2.7
django-cors-headers>=4.4
openai>=1.52.2
google-genai>=0.3.0

